requirements.txt
Django==1.5.4
Pillow==2.2.1
South==0.8.2
argparse==1.2.1
dj-database-url==0.2.2
dj-static==0.0.5
django-appconf==0.6
django-autocomplete-light==1.4.2
django-guardian==1.1.1
django-imagekit==3.0.3
django-registration==1.0
django-toolbelt==0.0.1
gunicorn==18.0
pilkit==1.1.5
psycopg2==2.5.1
six==1.4.1
static==0.4
wsgiref==0.1.2
